+++ CHANGELOG.md
@@ -160,9 +160,18 @@
 - check_squares_one, check_squares_multiple, eat_piece, pawn, knight are given the cell instead of searching for it
 
 
-
 # ADDED
 
 - New_piece on Cell
 
 - Print_board, print_turn on chess
+
+- on_board to Board
+
+# FIXED
+
+- Rook and Bishop movement
+
+# DELETED
+
+- on_board from Pieces